nd-video-file-stream: implement media controls
this implements pause/resume, seek and volume

if you think over 1000 lines of changes for such a simple feature is a bit extreme, you are not wrong.
since we are rendering stuff off-screen, we can't just rely on intervideo/audiosrc to give us all the controls. this makes controls very difficult since we now need to keep an internal state, and an external state, and keep them in sync
this is still a bit hacky and buggy around pause/resume, but i really can't think of any other way of going about this since just pausing the playbin is not enough to block playback

Signed-off-by: Bardia Moshiri <bardia@furilabs.com>